Stop haproxy service on standalone network nodes
Neutron uses haproxy for the metadata service and when the package is
installed the service will be automatically started on debian/ubuntu.
Neutron runs it's own haproxy instances and the systemd service is
not required to be running.

This patch stops the haproxy service for any neutron metadata hosts
which are not in the OSA haproxy_all ansible group.

Change-Id: I04d933007069e2c3164968b8add48db50210e25c