Make has_offline_migrations output predictable
Since we parse neutron-db-manage has_offline_migrations for specific
phrase in the output, we need to ensure that locale is correct.
As in case users locale is passed to the server (with SSH), this will
cause task failure.

Change-Id: Iec7f1314a2e4fa4256d9b24d3ea927bd5479e7a1
(cherry picked from commit 124dc769190bea27ca35e57d23615ba928a6af20)