L3 agent cleanup_on_shutdown
Once we do not kill keepalived for l3 agent, it might be usefull to
override that. This is possible with neutron_l3_cleanup_on_shutdown
When it set to True, keepalived will be restarted by l3 agent
except first service restart, where it wil be killed by handler,
since config should be loaded first.

Change-Id: I9eea72d68398f9fd272b1e9ae0c0c0198336c2f5