Retry applying OVN connection settings
This task runs immediately after one which may start the OVN
services and the unix socket files may not yet be present
when the command is run to configure the connection settings.

Introduce retires to the task to give time for the services to
start and the sockets to exist.

See https://paste.opendev.org/show/bPgVSIHyVPY5MwC373Zj/

Change-Id: I286169ca9ec493ef9ff1923249336cdc168619d0