Fix permissions for rootwrap files
At the moment we set 640 permissions to /etc/neutron/rootwrap.d folder.
While it doesn't cause any issues right now, since root still able to read files in there,
but this makes us to use root for services when it should not be needed.

Also playbook is not idempotent, as it changes permissions for same
directory multiple times during runtime.

Task for setting rootwrap permissions is removed, since it's behaviour is
weird by design of file module.
It can be applied only to directories, meaning that either directory will not
have execution permissions or all files inisde it will have executable flag.

Change-Id: I577221e94d6cf9d940ee310757383cee24b80a03