Remove handler delegation for DB contract
There is no reason to perform that kind of delegation, as handler should
run only for one host out of the play anyway.

Even more, this delegation might cause failures in case of running role
with limits, as `neutron_bin` on play host may not exist on the
delegated host during minor upgrades, for instance.

Change-Id: Ic8d8aae75dd58a30cd41327fe62009cc0dfb8bbb