Ensure proper permissions for OVN Metadata service
With change of the user under which ovn-metadata service is
running from root to neutron, it was clean forgot to change an
ownership for existing configuration produced be the service during
upgrades.

This patch adds an extra folder defenition that should ensure ownership
being correct for all files related to the ovn-metadata-proxy service.

Closes-Bug: #2077684
Change-Id: I8e82558fce8a420dca5fb5302dd5f73e40230e48