Define default value for `neutron_default_availability_zones`
OVN driver does have a different behaviour whenever scheduler hint
is provided or not [1]. In case it's not provided, it will filter out
all chassis which do not have AZ specified[2]. With that, we supply
an AZ for each chassis starting with [3].

So most painless way forward from here is to ensure that same hint will
be supplied for scheduling.

With that the change should be also safe for OVS/LXB drivers.

[1] https://opendev.org/openstack/neutron/src/commit/e9acc14d37c53b7ae25cf810ac74686023f25756/neutron/common/ovn/utils.py#L1071-L1078
[2] https://opendev.org/openstack/neutron/src/commit/e9acc14d37c53b7ae25cf810ac74686023f25756/neutron/common/ovn/utils.py#L896-L897
[3] https://review.opendev.org/c/openstack/openstack-ansible-os_neutron/+/904572/2/tasks/providers/setup_ovs_ovn.yml

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-openstack_hosts/+/930272
Change-Id: I6caa3f682133c89c1c997599043fe142c84eb016