Ensure that services that intended to stay disabled are not started
This mainly affects neutron-rpc-server service, which intended to stay
disabled/stopped.

While we've introduced logic in vars, which is respected by systemd role
it is not respected by the role handlers, which brings service back up
running unconditionally.

This actually breaks neutron behaviour after merging of [1], which relies
on neutron-rpc-server being disabled.

[1] https://review.opendev.org/c/openstack/openstack-ansible-os_neutron/+/927881

Change-Id: I28c928362ef009c1b49673005463b653d038faf9