Configure apparmor ovedrrides for haproxy
Despite haproxy apparmor profile is not loaded by default
and is part of extra-profiles which are not loaded by default,
thus not managed with aa-enforce or aa-disable.

Thus, there is no good reason to attempt unloading the profile,
as aa-disable will report that no profile is found even if it
will be loaded.

Instead, it's worth to place required overrides in case an
operator will manually load the profile, as such override
is respected by mentioned extra-profile.

Change-Id: I9028cff5c45ce6797742c99345d8dedc8c65b343