Ensure neutron_use_uwsgi is treated a bool
In case the value is passed as an extra arg, it may be treated as
a string, leading to an unexpected behavior.  For instnace.
`"{{ neutron_use_uwsgi | ternary('started', 'stopped') }}"` will result
in `started` if playbook is run with `-e neutron_use_uwsgi=false`

However, if value will be defined in a file, it will be respected.

Closes-Bug: #2096937
Change-Id: I25d517d7590a358ee15ddaa2d89b42f54207b518
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>