Remove group conditionals from venv_packages
neutron_venv_packages is used as a base for wheels build.
By default, wheels are being built against first host, which is usually
 a neutron_server.

In this case, wheels for these optional packages will not be build,
causing installation to fail.

We either need to have a different variables for wheels build and
package installation, or accept installing un-needed packages for
all venvs.

Change-Id: I973912f51f1846b1a8d0a2af5aa7a304a4c7a3a6
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>